requirements.txt
darkdetect==0.7.1
pillow==10.1.0
ttkbootstrap==1.10.1
#customtkinter==5.2.2
sv_ttk==2.6.0
